+++ src/pages/Home/Home.jsx
@@ -1,9 +1,9 @@
 import React from "react"
 import { motion } from "framer-motion"
 
-import Banner from "../components/Banner/Banner"
-import Projects from "../components/Projects/Projects"
-import Technologies from "../components/Technologies/Technologies"
+import HomeBanner from "./HomeBanner"
+// import Projects from "../../components/Projects/Projects"
+// import Technologies from "../../components/Technologies/Technologies"
 
 const Home = () => {
   return (
@@ -12,7 +12,7 @@ const Home = () => {
       animate={{ opacity: 1 }}
       exit={{ opacity: 0 }}
     >
-      <Banner />
+      <HomeBanner />
       {/* <Technologies />
       <Projects /> */}
     </motion.div>
