+++ src/pages/index.js
@@ -1,4 +1,4 @@
-import Home from "./Home"
+import Home from "./Home/Home"
 import AboutPage from "./AboutPage"
 import DetailPage from "./DetailPage"
